test_input/webpage/requirements.txt
flask==3.0.3
rembg==2.0.57
pillow==10.4.0
numpy==1.26.4
python-dotenv==1.0.1
flask-cors==4.0.1
gunicorn==22.0.0
requests==2.32.3
opencv-python==4.10.0.84
scikit-image==0.24.0
torch==2.3.1
torchvision==0.18.1
onnxruntime==1.18.1